+++ src/components/trips/ShowTripList.js
@@ -2,6 +2,7 @@ import React, { useEffect, useState } from "react";
 import axios from "axios";
 import { Link } from "react-router-dom";
 import TripCard from "./TripCard";
+import PrimaryBtn from "../common/PrimaryBtn";
 
 const ShowTripList = () => {
 
@@ -33,12 +34,7 @@ const ShowTripList = () => {
       <div>
         <div className="flex flex-col">
           <h1 className="text-center text-2xl my-4">My Trips</h1>
-          <Link
-            className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded text-center m-auto"
-            to="/createtrip"
-          >
-            Add a new Trip
-          </Link>
+          <PrimaryBtn path="/createtrip" name="Add a new Trip" />
         </div>
         <div className="flex">
             {tripList}
